Create images of a statMap

modified from oro.nifti:::image.nifti

# S3 method for statMap
image(x, thresh = 2.32, index = NULL,
  col = gray(0:64/64), colpos = redyellow(64), colneg = bluecyan(64),
  plane = c("axial", "coronal", "sagittal"), xlab = "", ylab = "",
  axes = FALSE, oma = rep(0, 4), mar = rep(0, 4), bg = "black",
  ...)

Arguments

x
the statMap object to display images of

thresh
A threshold to apply to the image, defaults to 2.32

index
Any selected image indexes to use for the z. defaults to NULL

col
a vector of colors to use for scaled intensities defaults to a grey scale.

colpos
a vector of colors to use for positive values.

colneg
a vector of colors to use for negative values.

plane
the plane to display, can be axial, coronal or sagittal.

xlab
a title for the x axis.

ylab
a title for the u axis.

axes
display axes, defaults to false

oma
A vector of the form c(bottom, left, top, right) giving the size of the outer margins in lines of text. Default to 0.

mar
A numerical vector of the form c(bottom, left, top, right) which gives the number of lines of margin to be specified on the four sides of the plot. Defaults to 0.

bg
background color, defaults to black.

...
additional arguments passed to par