Performs nonParametric Bootstrap Joint (nPBJ) Inference

Performs nonParametric Bootstrap Joint (nPBJ) Inference

npbj(images, form, formred, mask, data = NULL, W = NULL,
  template = NULL, nboot = 1000, statistic = npbj.sei, ...)

Arguments

images
A character vector of subjects' images or a 4D niftiImage object.

form
A formula describing the full model.

formred
A formula describing the reduced model.

mask
A character string of niftiImage object identifying voxels to be included in the analysis.

data
A data frame with the variables in form for associated with subject images.

W
A weight vector for the weighted least squares regression.

template
A template image that will be included in the returned statMap object.

nboot
Number of bootstraps to perform.

statistic
A function of the form statistic(stat, ...) that takes a vector stat and other arguments to compute an image summary statistic for each bootstrap. E.g. npbj.sei.

...
Arguments passed to statistic function.

Value

Returns a list of length length(cfts)+2. The first two elements contain statMap$stat and statMap$template. The remaining elements are lists containing the following:

pvalues

A vector of p-values corresponding to the cluster labels in clustermaps.

clustermap

A niftiImage object with the cluster labels.

pmap

A nifti object with each cluster assigned the negative log10 of its cluster extent FWE adjusted p-value.

CDF

A bootstrap CDF.

See also