this is an R wrapper function for PALM it has limited functionality for the features that PALM offers

this is an R wrapper function for PALM it has limited functionality for the features that PALM offers

palm(
  i,
  m,
  s = NULL,
  d,
  o,
  C = NULL,
  Cstat = "extent",
  tc = NULL,
  f = NULL,
  fonly = TRUE,
  n = 10000,
  eb = NULL,
  within = NULL,
  whole = NULL,
  ee = TRUE,
  T = FALSE,
  noniiclass = TRUE,
  logp = TRUE
)

Arguments

i

vector of strings pointing to input images or a single 4d nifti string.

m

mask file.

s

vector of strings pointing to input surfaces or a single 4d surface.

d

design matrix as a data frame or csv file.

o

output directory and prefix.

C

cluster forming threshold as a Z-value

Cstat

cluster statistic

tc

T-contrast matrix that randomise takes as matrix, data frame, or csv.

f

F-contrast matrix that randomise takes, as matrix data frame, or csv.

fonly

only run the F-tests.

n

number of permutations to run.

eb

exchangeability blocks (not functional).

within

not functional.

whole

not functional.

ee

exchangeable errors assumed.

T

Run TFCE.

noniiclass

Allow gzipped niftis.

logp

Compute log10p-values.

Value

Returns a string of the command that was evaluated.